Backend/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
openai
PyPDF2
azure-cosmos
azure-storage-blob
azure-cognitiveservices-speech